refactor(rtps_embedded): Widen DataSize_t to uint32_t (Phase 4, DATA_FRAG slice A)
Sample-size type DataSize_t was uint16_t, capping a whole sample at 64 KB and
silently truncating PayloadBuffer::spaceUsed() above that. Widen it to uint32_t
so a sample can exceed 64 KB internally (prerequisite for DATA_FRAG). Wire-format
neutral - only the internal type widens; on-the-wire length fields stay 16-bit:

- common/types.hpp: DataSize_t uint16_t -> uint32_t (the ~38 references are
  typed params/returns and widen automatically; spaceUsed() now holds >64 KB).
- MessageFactory.hpp: the one place a size feeds a 16-bit wire field
  (octetsToNextHeader for a non-fragmented DATA) is explicitly narrowed with
  static_cast<uint16_t> + comment (a single unfragmented DATA is <64 KB, so the
  narrowing is correct).
- rtps_participant: max_payload_size stays 65535 (one unfragmented DATA
  submessage is still bounded by the 16-bit octetsToNextHeader); the static_assert
  now checks against uint16_t max (raised when DATA_FRAG lands).

Gates: golden byte-identical (goldens never regenerated - proves wire neutrality
of the widening); engine + facade + typed loopbacks PASS; docker interop 8/8 PASS
vs ROS 2; esp32 example builds.

+++ components/rtps_embedded/include/rtps/common/types.hpp
@@ -39,7 +39,11 @@ namespace rtps {
 // TODO move types to where they are needed!
 
 typedef uint16_t Ip4Port_t;
-typedef uint16_t DataSize_t;
+// Internal whole-sample / payload size type. Widened to 32-bit so a single
+// sample can exceed 64 KB internally (prerequisite for DATA_FRAG). This is an
+// INTERNAL type only: on-the-wire per-submessage length fields (e.g.
+// SubmessageHeader::octetsToNextHeader) stay 16-bit per the RTPS spec.
+typedef uint32_t DataSize_t;
 typedef int8_t ParticipantId_t; // With UDP only 120 possible
 
 enum class EntityKind_t : uint8_t {
